Add ESLint and eslint-plugin-jsx-a11y dependencies and refactored signup
