+++ package.json
@@ -7,6 +7,8 @@
   },
   "devDependencies": {
     "@types/lodash": "^4.14.202",
-    "@types/lodash.debounce": "^4.0.9"
+    "@types/lodash.debounce": "^4.0.9",
+    "eslint": "^8.56.0",
+    "eslint-plugin-jsx-a11y": "^6.8.0"
   }
 }
